test: 자동차 이름 관련 테스트 추가
